refactor: enhance logging and error handling in PDF to Markdown conversion

+++ scripts/run_pdf2md.py
@@ -1,5 +1,6 @@
 #  Copyright (c) 2026. Programacion Cientifica, DISC, Antofagasta, Chile.
 import logging
+from logging import Logger
 from pathlib import Path
 
 import pymupdf4llm
@@ -17,16 +18,26 @@ def main():
     input_file = root_dir / "data" / "uso-de-python-para-la-modelacion-y-determinacion-de-vida-util-de-matrices-alimentarias.pdf"
     log.debug(f"input_file: {input_file}")
 
+    # the file don't exists
+    if not input_file.exists():
+        log.error(f"input file not found: {input_file}")
+        raise FileNotFoundError(f"input file not found: {input_file}")
+
     # output file (markdown)
     output_file = root_dir / "output" / "uso-de-python-para-la-modelacion-y-determinacion-de-vida-util-de-matrices-alimentarias.md"
     log.debug(f"output_file: {output_file}")
 
+    # create the directory if don't exists
+    output_file.parent.mkdir(parents=True, exist_ok=True)
+
     # read the input file into markdown
-    with benchmark("convert pdf2md", log):
+    with benchmark("pdf2md", log):
+        log.debug("converting pdf to markdown ..")
         md = pymupdf4llm.to_markdown(input_file)
 
     # write the markdown data
-    with benchmark("writting", log):
+    with benchmark("write", log):
+        log.debug(f"writing {len(md)} chars ..")
         output_file.write_text(md, encoding="utf-8")
 
 # Call the main function
@@ -34,9 +45,9 @@ if __name__ == '__main__':
     # configure the logging
     configure_logging(logging.DEBUG)
     # get the main logger
-    log = logging.getLogger(__name__)
+    log: Logger = logging.getLogger(__name__)
     # measure time
     with benchmark("main", log):
-        log.info("⬇️🏎️ starting ..")
+        log.info("️🏎️ starting ..")
         main()
-        log.info("⬆️🏁 done.")
+        log.info("️🏁 done.")
